requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-ninja==0.16.1
django-seed==0.3.1
Faker==10.0.0
psycopg2==2.9.2
pydantic==1.8.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
typing-extensions==4.0.1
tzdata==2021.5
PyJWT==2.0.1
